requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
aiogram>=3.0.0
skyfield>=1.45
pytz>=2023.3
python-dotenv>=1.0.0
reportlab>=4.0.0
pydantic>=2.0.0
jinja2>=3.0.0
pytest>=7.0.0
sentry-sdk>=1.0.0
